requirements.txt
# ML & Deep Learning
tensorflow>=2.19.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
pillow>=10.2.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Experiment Tracking
mlflow>=2.10.0

# Data Version Control
# Note: DVC may have compatibility issues with some pathspec versions
# Install separately if needed: pip install dvc
# dvc>=3.50.0
# pathspec>=0.12.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Monitoring & Logging
prometheus-fastapi-instrumentator>=6.1.0
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
